Fix bugs and warnings uncovered by type checker
The `*-query` functions, which produce query maps, are OK to annotate with
`:no-check` because we know they will satisfy the HMap type aliases, but the
use of `merge` means the type checker cannot _prove_ it conclusively.
If the user does happen to pass in extra keys that don't mean anything,
it won't be the end of the world.