Provide pure impl of a sheet-writer
The write fn having its own thread pool is silly b/c an app generally
would like to have control over the total number of concurrent threads
talking to an api. This version express the write operation as a seq of
seq of thunks and provides an execute! fn to invoke them on a thread
pool.